+++ ydb/tests/functional/audit/ya.make
@@ -14,6 +14,7 @@ INCLUDE(${ARCADIA_ROOT}/ydb/tests/ydbd_dep.inc)
 
 TEST_SRCS(
     conftest.py
+    helpers.py
     test_auditlog.py
     test_canonical_records.py
 )
